backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
openai>=1.6.1,<2.0.0
google-generativeai==0.3.1
chromadb==0.4.18
pymupdf==1.23.8
serpapi==0.1.5
python-dotenv==1.0.0
httpx==0.25.2
numpy==1.24.3
langchain==0.0.350
langchain-openai==0.0.2
langchain-google-genai==0.0.6
cohere>=4.37,<5.0.0